ACFG04/ingeniousLib/requirements_3_10.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
paho-mqtt==1.6.1
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
tzdata==2023.4
urllib3==2.2.0
